refactor(cli): extract config resolvers
